move cloudstack template builder to it's own file
